+++ src/front/js/component/princing.js
@@ -3,8 +3,9 @@ import { Link } from "react-router-dom";
 const Pricing =(props)=>{
     return(
         <div className="container">
+          
         <div className="card mb-4 box-shadow">
-          <div className="card-header">
+          <div id="card-header" className="card-header">
             <h4 className="my-0 font-weight-normal">{props.planes}</h4>
           </div>
           <div className="card-body">
