+++ src/front/js/pages/galeria.js
@@ -1,0 +1,22 @@
+import React from "react";
+import React from "react"
+import "../../styles/home.css";
+import basket from "../../img/basket7.jpeg";
+
+import basket5 from "../../img/basket5.jpeg";
+import basket12 from "../../img/basket12.jpg";
+import basket13 from "../../img/basket13.jpg";
+import basket14 from "../../img/basket14.jpeg";
+import basket15 from "../../img/basket15.jpg";
+import basket16 from "../../img/basket16.jpg";
+import lebron17 from "../../img/lebron17.jpg";
+import basket6 from "../../img/basket6.jpeg";
+const Galeria=()=>{
+    return(
+
+      <div className="App">
+      <Galeria/>
+     </div>
+    );
+}
+export default Galeria;
